Fix ILVerify build and simplify now that these concepts are in the base type system.
